HAWQ-1228. Moved ProfileFactory to pxf-hive.
